samv71 pmc: add generic and programmable clocks, utmi-pll
Basically clocks with multiple sources and their own divider; very
similar to each other, but named separately in the documentation.

The utmi pll provides a 480MHz clock for the HighSpeed USB UTMI
that can then be divided down to e.g. 8MHz for CAN, while
the main 300MHz clock cannot. Downside of the utmi pll is
the limited multipliers: 30 and 40, limiting the input clock
to 12 or 16 MHz.